requirements.txt
-i https://pypi.org/simple
aiosqlite==0.17.0 ; python_version >= '3.6'
annotated-types==0.5.0 ; python_version >= '3.7'
anyio==3.7.1 ; python_version >= '3.7'
click==8.1.7 ; python_version >= '3.7'
fastapi==0.103.1
feedgen==0.9.0
h11==0.14.0 ; python_version >= '3.7'
httptools==0.6.0
idna==3.4 ; python_version >= '3.5'
iso8601==1.1.0 ; python_full_version >= '3.6.2' and python_version < '4.0'
lxml==4.9.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==2.3.0 ; python_version >= '3.7'
pydantic-core==2.6.3 ; python_version >= '3.7'
pypika-tortoise==0.1.6 ; python_version >= '3.7' and python_version < '4.0'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.0
pytz==2023.3
pyyaml==6.0.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
tortoise-orm==0.20.0
typing-extensions==4.7.1 ; python_version >= '3.7'
uvicorn[standard]==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
